firmware/server/live2d/requirements.txt
paho-mqtt>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
PyQt5>=5.15.0
PyOpenGL>=3.1.6
live2d-py>=0.5.3.5
